+++ src/main/java/io/github/astrapi69/gradle/migration/info/ProjectTomlStructureInfo.java
@@ -22,29 +22,23 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.gradle.migration.data;
+package io.github.astrapi69.gradle.migration.info;
 
 import java.io.File;
 
+import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Data;
-import lombok.experimental.SuperBuilder;
+import lombok.experimental.FieldDefaults;
 
 @Data
-@SuperBuilder
-public class CopyGradleRunConfigurations
+@Builder
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class ProjectTomlStructureInfo
 {
-	public static final String IDEA_DIR_NAME = ".idea";
-	public static final String RUN_CONFIGURATIONS_DIR_NAME = "runConfigurations";
-	File ideaSourceDir;
-	File ideaTargetDir;
-	String sourceFilenamePrefix;
-	File sourceProjectDir;
+	File gradleDirectory;
+	String sourceProjectDirNamePrefix;
+	String targetProjectDirNamePrefix;
 	String sourceProjectName;
-	File sourceRunConfigDir;
-	String targetFilenamePrefix;
-	File targetProjectDir;
-	boolean onlyRunConfigurations;
-	boolean runConfigurationsInSameFolder;
 	String targetProjectName;
-	File targetRunConfigDir;
 }