+++ src/main/java/io/github/astrapi69/gradle/migration/runner/GradleRunConfigurationsCopier.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.gradle.migration.data;
+package io.github.astrapi69.gradle.migration.runner;
 
 import java.io.File;
 import java.io.IOException;
@@ -49,6 +49,9 @@ import io.github.astrapi69.file.read.ReadFileExtensions;
 import io.github.astrapi69.file.rename.RenameFileExtensions;
 import io.github.astrapi69.file.search.FileSearchExtensions;
 import io.github.astrapi69.file.write.StoreFileExtensions;
+import io.github.astrapi69.gradle.migration.extension.DependenciesExtensions;
+import io.github.astrapi69.gradle.migration.info.CopyGradleRunConfigurations;
+import io.github.astrapi69.gradle.migration.info.DependenciesInfo;
 import io.github.astrapi69.io.StreamExtensions;
 import io.github.astrapi69.string.StringExtensions;
 
@@ -56,6 +59,27 @@ public class GradleRunConfigurationsCopier
 {
 	private final CopyGradleRunConfigurations copyGradleRunConfigurations;
 	private final Logger log = Logger.getLogger(GradleRunConfigurationsCopier.class.getName());
+
+	public static void copyOnlyRunConfigurations(String sourceProjectName, String targetProjectName,
+		String sourceProjectDirNamePrefix, String targetProjectDirNamePrefix)
+	{
+
+		CopyGradleRunConfigurations copyGradleRunConfigurationsData = GradleRunConfigurationsCopier
+			.newCopyGradleRunConfigurations(sourceProjectName, targetProjectName,
+				sourceProjectDirNamePrefix, targetProjectDirNamePrefix, true, false);
+		GradleRunConfigurationsCopier.of(copyGradleRunConfigurationsData).copy();
+	}
+
+	public static void copyRunConfigurations(String sourceProjectName, String targetProjectName,
+		String sourceProjectDirNamePrefix, String targetProjectDirNamePrefix,
+		boolean onlyRunConfigurations, boolean runConfigurationsInSameFolder)
+	{
+
+		CopyGradleRunConfigurations copyGradleRunConfigurationsData = GradleRunConfigurationsCopier
+			.newCopyGradleRunConfigurations(sourceProjectName, targetProjectName,
+				sourceProjectDirNamePrefix, targetProjectDirNamePrefix, true, false);
+		GradleRunConfigurationsCopier.of(copyGradleRunConfigurationsData).copy();
+	}
 
 	private GradleRunConfigurationsCopier(CopyGradleRunConfigurations copyGradleRunConfigurations)
 	{
@@ -101,6 +125,7 @@ public class GradleRunConfigurationsCopier
 			.runConfigurationsInSameFolder(runConfigurationsInSameFolder).build();
 	}
 
+	@Deprecated
 	public static String getProjectVersionKeyName(String projectName)
 	{
 		String camelCased = WordUtils.capitalizeFully(projectName, new char[] { '-' })
@@ -196,7 +221,7 @@ public class GradleRunConfigurationsCopier
 		File buildGradle = new File(targetProjectDir, DependenciesInfo.BUILD_GRADLE_NAME);
 		File gradleProperties = new File(targetProjectDir, DependenciesInfo.GRADLE_PROPERTIES_NAME);
 		FileFactory.newFile(gradleProperties);
-		String dependenciesContent = getDependenciesContent(buildGradle);
+		String dependenciesContent = DependenciesExtensions.getDependenciesContent(buildGradle);
 		List<String> stringList = getDependenciesAsStringList(dependenciesContent);
 		DependenciesInfo dependenciesInfo = getGradlePropertiesWithVersions(stringList);
 		String newDependenciesContent = getNewDependenciesContent(dependenciesInfo);
@@ -216,12 +241,47 @@ public class GradleRunConfigurationsCopier
 		return stringList;
 	}
 
-	public String getDependenciesContent(File buildGradle) throws IOException
+	public static String getContentOf(String section, File buildGradle) throws IOException
 	{
 		String buildGradleContent = ReadFileExtensions.fromFile(buildGradle);
-		int indexOfStart = buildGradleContent.indexOf("dependencies {");
-		int indexOfEnd = buildGradleContent.substring(indexOfStart).indexOf("}") + indexOfStart + 1;
-		return buildGradleContent.substring(indexOfStart, indexOfEnd);
+		return getContentOf(section, buildGradleContent);
+	}
+
+	public static String getContentOf(String section, String buildGradleContent)
+	{
+		int indexOfStart = buildGradleContent.indexOf(section + " {");
+		if (indexOfStart == -1)
+		{
+			throw new IllegalArgumentException("Section not found: " + section);
+		}
+		int indexOfEnd = findClosingBrace(buildGradleContent, indexOfStart + section.length() + 2);
+		return buildGradleContent.substring(indexOfStart, indexOfEnd + 1);
+	}
+
+	private static int findClosingBrace(String content, int start)
+	{
+		int braceCount = 1;
+		for (int i = start; i < content.length(); i++)
+		{
+			if (content.charAt(i) == '{')
+			{
+				braceCount++;
+			}
+			else if (content.charAt(i) == '}')
+			{
+				braceCount--;
+				if (braceCount == 0)
+				{
+					return i;
+				}
+			}
+		}
+		throw new IllegalArgumentException("No matching closing brace found.");
+	}
+
+	public String getBuildscriptContent(File buildGradle) throws IOException
+	{
+		return GradleRunConfigurationsCopier.getContentOf("buildscript", buildGradle);
 	}
 
 	private DependenciesInfo getGradlePropertiesWithVersions(List<String> stringList)