+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 ## Change log
 ----------------------
 
+Version 2.2
+-------------
+
+ADDED:
+
+- new class MigrateToTomlVersions that provides methods for migrate to new gradle project with libs.versions.toml file
+- new libs.versions.toml file for new automatic catalog versions update
+- new method for get the content of a section from the build.gradle file
+
+CHANGED:
+
+- update gradle to new version 8.8
+- update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new minor version 7.0.0.BETA1
+- update of test dependency mockito-core to new version 5.12.0
+- update of test dependency junit-jupiter-api to new version 5.11.0-M2
+- rename of package data to info
+
 Version 2.1
 -------------